% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise2.R
\name{praise2}
\alias{praise2}
\title{Deliver praise}
\usage{
praise2(name, punctuation = "!")
}
\arguments{
\item{name}{text string; This is the name of the person I want to praise}

\item{punctuation}{text string; This is our emphasis as a text input.}
}
\value{
Text string with praise
}
\description{
This function is super useful when you're feeling sad
and is useful for everyone.
}
\examples{
praise2(name = "Steph", punctuation = "!")
}
